Первый рефакторинг программы под SOLID
