+++ src/main/java/org/example/entity/Task.java
@@ -1,15 +1,12 @@
-package org.example;
+package org.example.entity;
 
 import lombok.Data;
 
 @Data
 public class Task {
-
-    // поля класса
     private boolean done;
     private String description;
 
-    // конструкторы
     public Task() { }
 
     public Task(String description) {
