+++ src/main/java/org/example/userCommand/interfaceCommand/CommandPrint.java
@@ -1,0 +1,10 @@
+package org.example.userCommand.interfaceCommand;
+
+import org.example.entity.Task;
+
+import java.util.Map;
+import java.util.Scanner;
+
+public interface CommandPrint {
+    void print(Scanner scanner);
+}
